refactor: 調整資料路徑與模組引用，修正 UI 模組路徑錯誤
- 統一使用 core.config 管理 DATA_DIR、MODEL_DIR
- 修正 AIPredictor 使用相對路徑導致 ui/data/models 的錯誤問題
- 更新 betting_page、ai_control 等頁面引用方式
- UI 初始化測試通過
