+++ main.py
@@ -1,4 +1,5 @@
 from ui.main_window import MainWindow
+
 from PySide6.QtWidgets import QApplication
 import sys
